Create a new 'ots-distribution' module that provides a deduplicated uber JAR for OTS
With the introduction of the shade maven plugin (fat jars), the module JARs
also include all dependencies of each module. Because each OTS module is built
independently, this means that the dependencies are duplicated across all
modules and the combined module size exceeds 100MB. As PyPi limits distributions
to 100MB, OTS cannot be distributed on PyPi with the current setup.

Therefore, this adds a separate 'ots-distribution' which combines all OTS
modules and their dependencies. With the deduplication applied,
a distribution size of roughly 15MB is achieved.